+++ foxtive-ntex/src/error.rs
@@ -1,6 +1,9 @@
-use crate::http::response::anyhow::helpers::{make_response, make_status_code};
+use crate::error::helpers::make_http_error_response;
+use crate::http::response::anyhow::helpers::make_status_code;
 use foxtive::prelude::AppMessage;
 use foxtive::Error;
+#[cfg(feature = "multipart")]
+use foxtive_ntex_multipart::{ErrorMessage as MultipartErrorMessage, MultipartError};
 use ntex::http::error::PayloadError;
 use ntex::http::StatusCode;
 use ntex::web::error::BlockingError;
@@ -25,7 +28,7 @@ pub enum HttpError {
     ValidationError(#[from] validator::ValidationErrors),
     #[cfg(feature = "multipart")]
     #[error("Multipart Error: {0}")]
-    MultipartError(#[from] foxtive_ntex_multipart::MultipartError),
+    MultipartError(#[from] MultipartError),
 }
 
 impl HttpError {
@@ -54,15 +57,112 @@ impl WebResponseError for HttpError {
         match self {
             HttpError::AppMessage(m) => m.status_code(),
             HttpError::AppError(e) => make_status_code(e),
+            #[cfg(feature = "validator")]
+            HttpError::ValidationError(_) => StatusCode::BAD_REQUEST,
+            HttpError::PayloadError(_) => StatusCode::BAD_REQUEST,
+            #[cfg(feature = "multipart")]
+            HttpError::MultipartError(err) => match err {
+                MultipartError::ValidationError(err) => match err.error {
+                    MultipartErrorMessage::InvalidFileExtension(_)
+                    | MultipartErrorMessage::InvalidContentType(_) => {
+                        StatusCode::UNSUPPORTED_MEDIA_TYPE
+                    }
+                    _ => StatusCode::BAD_REQUEST,
+                },
+                _ => StatusCode::BAD_REQUEST,
+            },
             _ => StatusCode::INTERNAL_SERVER_ERROR,
         }
     }
 
     fn error_response(&self, _: &HttpRequest) -> HttpResponse {
-        match self {
+        make_http_error_response(self)
+    }
+}
+
+mod helpers {
+    use crate::enums::ResponseCode;
+    use crate::helpers::responder::Responder;
+    use crate::http::response::anyhow::helpers::make_response;
+    use crate::http::HttpError;
+    use foxtive::prelude::AppMessage;
+    use ntex::web::HttpResponse;
+
+    pub(crate) fn make_http_error_response(err: &HttpError) -> HttpResponse {
+        match err {
             HttpError::AppMessage(m) => make_response(&m.clone().ae()),
             HttpError::AppError(e) => make_response(e),
+            #[cfg(feature = "validator")]
+            HttpError::ValidationError(e) => {
+                Responder::send_msg(e.errors(), ResponseCode::BadRequest, "Validation Error")
+            }
+            HttpError::PayloadError(e) => {
+                Responder::send_msg(e.to_string(), ResponseCode::BadRequest, "Payload Error")
+            }
+            #[cfg(feature = "multipart")]
+            HttpError::MultipartError(err) => Responder::send_msg(
+                err.to_string(),
+                ResponseCode::BadRequest,
+                "File Upload Error",
+            ),
             _ => make_response(&foxtive::Error::from(AppMessage::InternalServerError)),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use foxtive::Error;
+
+    #[test]
+    fn test_app_error() {
+        let error = HttpError::AppError(Error::from(AppMessage::InternalServerError));
+        let app_error = make_http_error_response(&error);
+        assert_eq!(app_error.status(), 500);
+    }
+
+    #[test]
+    fn test_app_message() {
+        let error = HttpError::AppMessage(AppMessage::InternalServerError);
+        let app_error = make_http_error_response(&error);
+        assert_eq!(app_error.status(), 500);
+    }
+
+    #[test]
+    fn test_std_error() {
+        let error = HttpError::Std(Box::new(std::io::Error::new(std::io::ErrorKind::Other, "Test")));
+        let app_error = make_http_error_response(&error);
+        assert_eq!(app_error.status(), 500);
+    }
+
+    #[test]
+    fn test_payload_error() {
+        let error = HttpError::PayloadError(PayloadError::Overflow);
+        let app_error = make_http_error_response(&error);
+        assert_eq!(app_error.status(), 400);
+    }
+
+    #[cfg(feature = "validator")]
+    #[test]
+    fn test_validation_error() {
+        let error = HttpError::ValidationError(validator::ValidationErrors::new());
+        let app_error = make_http_error_response(&error);
+        assert_eq!(app_error.status(), 400);
+    }
+
+    #[cfg(feature = "multipart")]
+    #[test]
+    fn test_multipart_error() {
+        use foxtive_ntex_multipart::InputError;
+
+        let error = HttpError::MultipartError(MultipartError::ValidationError(InputError {
+            error: MultipartErrorMessage::InvalidFileExtension(Some("mp4".to_string())),
+            name: "image".to_string(),
+        }));
+
+        let app_error = make_http_error_response(&error);
+
+        assert_eq!(app_error.status(), 400);
     }
 }